🎨 move out server component
