+++ src/api/notice/notice.ts
@@ -1,5 +1,4 @@
 import dayjs from 'dayjs';
-import { cookies } from 'next/headers';
 
 import { gql } from '@/generated';
 
@@ -48,26 +47,6 @@ export interface Notices {
   total: number;
 }
 
-export const getAllNotices = async (
-  params: NoticePaginationParams & NoticeSearchParams = {},
-) => {
-  const cookieStore = cookies();
-  return api
-    .get<Notices>('/notice', {
-      params,
-      headers: {
-        Authorization: `Bearer ${cookieStore.get('access_token')}`,
-      },
-    })
-    .then(({ data }) => ({
-      ...data,
-      list: data.list.map(({ imageUrl, ...notice }) => ({
-        ...notice,
-        deadline: notice.deadline ? notice.deadline : null,
-        imageUrl: imageUrl ? imageUrl : null,
-      })),
-    }));
-};
 export const getNotice = async (id: number) =>
   api.get<NoticeDetail>(`/notice/${id}`).then(({ data }) => ({
     ...data,
